requirements.txt
feedparser>=6.0
trafilatura>=2.0
httpx>=0.27
pyyaml>=6.0
rich>=13.0
anthropic>=0.40
flask>=3.0
gunicorn>=23.0
beautifulsoup4>=4.12
lxml>=5.0
google-genai>=1.0
openai>=1.0
